libs/python/mf-nacos-service-registrar/pyproject.toml
[project]
name = "mf-nacos-service-registrar"
version = "0.1.0"
description = "Python Nacos 服务注册/发现二方包，适用于 MF 生态"
authors = [
    {name = "Your Name", email = "your@email.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "nacos-sdk-python>=2.0.0"
]

[tool.setuptools.packages.find]
where = ["src"] 